versao final

+++ src/app/Repositories/ingredientsGroup/deleteIngredientsGroups.ts
@@ -7,6 +7,17 @@ export const deleteIngredientsGroups = async (req: Request, res: Response) => {
   try {
     const { ingredientId } = req.params;
 
+    const existIngredientBond = await Product.findOne({
+      "ingredients.ingredient": ingredientId,
+    });
+
+
+    if(existIngredientBond){
+      return res.status(400).json({
+        error: true,
+        message: "Ocorreu um erro ao tentar deletar um ingrediente, pois ele esta vinculado á um grupo.",
+      });
+    }
     await IngredientGroup.findByIdAndDelete(ingredientId);
 
     res.status(200).json({ deleted: true });
